+++ src/Utils/index.ts
@@ -77,3 +77,16 @@ export const params = (a: any) => {
   output = s.join('&').replace(r20, '+');
   return output;
 };
+
+export const buildURLwithQueryParams = (urlString: string, params: any) => {
+  const {
+    url,
+    url: { searchParams }
+  } = { url: new URL(urlString) };
+
+  Object.keys(params).forEach(key => {
+    searchParams.set(key, params[key]);
+  });
+
+  return url.href;
+};