+++ src/index.d.ts
@@ -346,17 +346,17 @@ export declare class DLC {
 }
 
 export interface CreatePaymentData {
-  number?: number | string;
-  cardName?: string;
-  expMonth?: number;
-  expYear?: number;
-  cvv?: number;
-  accessFee?: number;
-  paymentMethod?: string;
-  referrer?: string;
+  number: number | string;
+  cardName: string;
+  expMonth: number;
+  expYear: number;
+  cvv: number;
+  accessFee: number;
+  paymentMethod: string;
+  referrer: string;
+  returnUrl: string;
   voucherCode?: string;
   brandingId?: number;
-  returnUrl?: string;
 }
 
 export interface PayPalParamsData {
@@ -539,6 +539,7 @@ export interface CreateSubscriptionData {
   referrer: string;
   voucherCode?: string;
   brandingId?: number;
+  returnUrl: string;
 }
 
 export declare interface SubscriptionDetails {