+++ src/lib/Subscription.ts
@@ -99,20 +99,22 @@ class Subscription {
    *  referrer: string
    *  voucherCode?: string
    *  brandingId?: number
+   *  returnUrl?: string
    * }
    * @example
    *     InPlayer.Subscription
    *     .create({
-   *        number: 1,
-   *        cardName: 'Payoneer',
-   *        expMonth: 11,
-   *        expYear: 12,
-   *        cvv: 546,
-   *        accessFee: 13.4,
-   *        paymentMethod: 'card',
-   *        referrer: 'http://localhost:3000',
-   *        voucherCode: '123123125914i2erjfg'
-   *        brandingId?: 1234
+   *          number: 1,
+   *          cardName: 'Payoneer',
+   *          expMonth: 11,
+   *          expYear: 12,
+   *          cvv: 546,
+   *          accessFee: 13.4,
+   *          paymentMethod: 'card',
+   *          referrer: 'http://localhost:3000',
+   *          voucherCode: '123123125914i2erjfg',
+   *          brandingId?: 1234,
+   *          returnUrl?: 'https://event.inplayer.com/staging',
    *        }
    *     )
    *     .then(data => console.log(data));
@@ -129,6 +131,7 @@ class Subscription {
       payment_method: data.paymentMethod,
       referrer: data.referrer,
       branding_id: data.brandingId,
+      returnUrl: data.returnUrl,
     };
 
     if (data.voucherCode) {